+++ .gitignore
@@ -126,3 +126,4 @@ dmypy.json
 
 *.mp4
 client.py
+wsgi.py
