+++ __init__.py
@@ -1,1 +1,22 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from dotenv import load_dotenv
+import os
 
+db = SQLAlchemy()
+
+
+def init_app():
+    app = Flask(__name__, instance_relative_config=False)
+    load_dotenv()
+    app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('DB_URI')
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
+    db.init_app(app)
+    os.makedirs('videos', exist_ok=True)
+
+    with app.app_context():
+        from . import routes
+        db.create_all()
+
+        return app
